Title: 최대 수 만들기, Time: 192ms, Memory: 2MB, Status: Passed - CodeTree

+++ 250227/최대 수 만들기/make-biggest-num.cpp
@@ -7,7 +7,7 @@ bool cmp(string left, string right) {
     string left_right = left + right;
     string right_left = right + left;
 
-    if (stoi(left_right) > stoi(right_left))
+    if (stol(left_right) > stol(right_left))
         return true;
     return false;
 }
